+++ CHANGELOG.md
@@ -1,9 +1,19 @@
-# proteusPy Changelog
+# proteusPy ChangeLog
 
 Notable changes to the ``proteusPy`` project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [v0.98] - 2024-12-24
+
+### Added
+
+- Dynamic resolution for the rcsb_viewer List view.
+
+### Changed
+
+- ``DisulfideList`` code optimization
 
 ## [V0.97.17] - 2024-11-30